+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `_stackPushVariable(variable)` and `_isIndirectVariable(variable)` helpers for use in plugins
 - Add "View Script Uses" feature accessible from the script context menu, allowing you to see all places where a script has been used in your project
 - Add ability to change compiler optimisation settings in Build Options section on Settings page [@patrickmollohan](https://github.com/patrickmollohan)
+- Add event "Set Actor Collision Bounding Box" to modify collision shape
 
 ### Changed